// src/models/public/ActionReportValues.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { AccountsAccountId } from './Accounts';
import type { ActionReportsActionReportId } from './ActionReports';
import type { UnitsUnitId } from './Units';

/** Identifier type for public.action_report_values */
export type ActionReportValuesActionReportValueId = string & { __brand: 'public.action_report_values' };

/**
 * Represents the table public.action_report_values
 * value-ing the success of actions
 */
export default interface ActionReportValues {
  action_report_value_id: ActionReportValuesActionReportValueId;

  /** redundant account_id enhances data safety */
  account_id: AccountsAccountId | null;

  action_report_id: ActionReportsActionReportId | null;

  unit_id: UnitsUnitId | null;

  /** Used for integer values */
  value_integer: number | null;

  /** Used for numeric values */
  value_numeric: number | null;

  /** Used for text values */
  value_text: string | null;

  label: string | null;

  created_at: Date;

  updated_at: Date;

  updated_by: string | null;
}

/**
 * Represents the initializer for the table public.action_report_values
 * value-ing the success of actions
 */
export interface ActionReportValuesInitializer {
  /** Default value: uuid_generate_v7() */
  action_report_value_id?: ActionReportValuesActionReportValueId;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  action_report_id?: ActionReportsActionReportId | null;

  unit_id?: UnitsUnitId | null;

  /** Used for integer values */
  value_integer?: number | null;

  /** Used for numeric values */
  value_numeric?: number | null;

  /** Used for text values */
  value_text?: string | null;

  label?: string | null;

  /** Default value: now() */
  created_at?: Date;

  /** Default value: now() */
  updated_at?: Date;

  updated_by?: string | null;
}

/**
 * Represents the mutator for the table public.action_report_values
 * value-ing the success of actions
 */
export interface ActionReportValuesMutator {
  action_report_value_id?: ActionReportValuesActionReportValueId;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  action_report_id?: ActionReportsActionReportId | null;

  unit_id?: UnitsUnitId | null;

  /** Used for integer values */
  value_integer?: number | null;

  /** Used for numeric values */
  value_numeric?: number | null;

  /** Used for text values */
  value_text?: string | null;

  label?: string | null;

  created_at?: Date;

  updated_at?: Date;

  updated_by?: string | null;
}